refactor: extract accessioning logic out of sample model into a concern
